backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
PyGithub==2.1.1
requests==2.31.0
python-dotenv==1.0.0
openai>=1.12.0
